move test to new module
